+++ libs/spark/CHANGELOG.md
@@ -4,7 +4,11 @@
 
 ### Fixes
 
+- **Avatar**
+  - custom classes not being global.
 - **Tag**
+  - custom classes not being global.
+- **Typography**
   - custom classes not being global.
 
 # [v0.11.1](https://github.com/prenda-school/prenda-spark/compare/v0.11.0...v0.11.1) (2021-09-07)